Break create-fills job into smaller modules
