+++ src/jobs/create-fills/create-fill.js
@@ -1,95 +1,15 @@
-const { assetDataUtils } = require('@0x/order-utils');
-
 const _ = require('lodash');
 
-const {
-  MissingBlockError,
-  UnsupportedAssetError,
-  UnsupportedProtocolError,
-} = require('../../errors');
-const Event = require('../../model/event');
-const Fill = require('../../model/fill');
+const { MissingBlockError } = require('../../errors');
 const getBlock = require('../../util/ethereum/get-block');
 const getRelayerForFill = require('../../fills/get-relayer-for-fill');
 const getRoundedDates = require('./get-rounded-dates');
+const normalizeFillArgs = require('./normalize-fill-args');
 const tokenCache = require('../../tokens/token-cache');
-
-const decodeAssetData = encodedData => {
-  const {
-    decodeAssetDataOrThrow,
-    isERC20AssetData,
-    isERC721AssetData,
-  } = assetDataUtils;
-
-  const assetData = decodeAssetDataOrThrow(encodedData);
-
-  if (isERC20AssetData(assetData)) {
-    return assetData;
-  }
-
-  if (isERC721AssetData(assetData)) {
-    return {
-      ...assetData,
-      tokenId: assetData.tokenId.toNumber(),
-    };
-  }
-
-  return null; // Unsupported asset type
-};
-
-const normaliseFillArguments = (args, protocolVersion) => {
-  if (_.isUndefined(protocolVersion) || protocolVersion === 1) {
-    return _.pick(
-      args,
-      'feeRecipient',
-      'filledMakerTokenAmount',
-      'filledTakerTokenAmount',
-      'maker',
-      'makerToken',
-      'orderHash',
-      'paidMakerFee',
-      'paidTakerFee',
-      'taker',
-      'takerToken',
-    );
-  }
-
-  if (protocolVersion === 2) {
-    const makerAsset = decodeAssetData(args.makerAssetData);
-    const takerAsset = decodeAssetData(args.takerAssetData);
-
-    if (makerAsset === null) {
-      throw new UnsupportedAssetError(`Event has unsupported maker asset`);
-    }
-
-    if (takerAsset === null) {
-      throw new UnsupportedAssetError(`Event has unsupported taker asset`);
-    }
-
-    return {
-      feeRecipient: args.feeRecipientAddress,
-      filledMakerTokenAmount: args.makerAssetFilledAmount,
-      filledTakerTokenAmount: args.takerAssetFilledAmount,
-      maker: args.makerAddress,
-      makerAsset,
-      makerToken: _.get(makerAsset, 'tokenAddress'), // TODO: Deprecate in favor of makerAsset
-      orderHash: args.orderHash,
-      paidMakerFee: args.makerFeePaid,
-      paidTakerFee: args.takerFeePaid,
-      senderAddress: args.senderAddress,
-      taker: args.takerAddress,
-      takerAsset,
-      takerToken: _.get(takerAsset, 'tokenAddress'), // TODO: Deprecate in favor of takerAsset
-    };
-  }
-
-  throw new UnsupportedProtocolError(
-    `Event has unrecognised protocol version: ${protocolVersion}`,
-  );
-};
 
 const createFill = async event => {
   const { protocolVersion } = event;
+
   const {
     args,
     blockHash,
@@ -97,6 +17,7 @@ const createFill = async event => {
     logIndex,
     transactionHash,
   } = event.data;
+
   const {
     feeRecipient,
     filledMakerTokenAmount,
@@ -111,7 +32,7 @@ const createFill = async event => {
     taker,
     takerAsset,
     takerToken,
-  } = normaliseFillArguments(args, protocolVersion);
+  } = normalizeFillArgs(args, protocolVersion);
 
   const block = await getBlock(blockHash);
 
@@ -154,8 +75,7 @@ const createFill = async event => {
     transactionHash,
   };
 
-  await Fill.create(fill);
-  await Event.updateOne({ _id: event._id }, { fillCreated: true });
+  return fill;
 };
 
 module.exports = createFill;
